migrating-to-python3/service-account-scopes/requirements.txt
Flask
google-auth
google-cloud-storage
requests
